Extract backup storage helpers
