+++ test/lib/server-backup.test.ts
@@ -3,7 +3,8 @@ import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
 import { zipSync } from "fflate";
 import { handleRequest } from "#routes";
-import { backupDir, createBackupZip } from "#shared/db/backup.ts";
+import { createBackupZip } from "#shared/db/backup.ts";
+import { backupDir } from "#shared/db/backup-storage.ts";
 import { getDb } from "#shared/db/client.ts";
 import { invalidateInitDbCache } from "#shared/db/migrations.ts";
 import { downloadRaw, uploadRaw } from "#shared/storage.ts";
